Restructure the library and formulas
Currently, there are library code and formulas are hanging around
everywhere, this is blocking us from using some ruby specific meta
programming to load these dynamically.

This commit move out all library related code to the main fontist
module, and leave out the formulas module for actual formulas. This
also migrate all existing formulas to use this new structure.
