+++ lib/fontist/utils.rb
@@ -1,0 +1,8 @@
+require "fontist/utils/dsl"
+require "fontist/utils/zip_extractor"
+require "fontist/utils/exe_extractor"
+
+module Fontist
+  module Utils
+  end
+end
